+++ src/main/java/com/example/helloboot/springBoot/binlogListener/binlog/BinlogEventV4Header.java
@@ -4,7 +4,7 @@ public interface BinlogEventV4Header {
 
     int getHeaderLength();
 
-    int getPosition();
+    long getPosition();
 
     long getTimeStamp();
